+++ extensions/yii/gii/generators/crud/templates/controller.php
@@ -19,7 +19,7 @@ if ($modelClass === $searchModelClass) {
 
 /** @var ActiveRecordInterface $class */
 $class = $generator->modelClass;
-$pks = $class::primaryKey();
+$pks = (array) $class::primaryKey();
 $urlParams = $generator->generateUrlParams();
 $actionParams = $generator->generateActionParams();
 $actionParamComments = $generator->generateActionParamComments();