+++ tests/unit/data/ar/elasticsearch/Order.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace yiiunit\data\ar\elasticsearch;
+use yii\elasticsearch\Command;
 
 /**
  * Class Order
@@ -12,24 +13,29 @@ namespace yiiunit\data\ar\elasticsearch;
  */
 class Order extends ActiveRecord
 {
+	public static function primaryKey()
+	{
+		return 'id';
+	}
+
 	public function attributes()
 	{
-		return ['customer_id', 'create_time', 'total'];
+		return ['id', 'customer_id', 'create_time', 'total'];
 	}
 
 	public function getCustomer()
 	{
-		return $this->hasOne(Customer::className(), [ActiveRecord::PRIMARY_KEY_NAME => 'customer_id']);
+		return $this->hasOne(Customer::className(), ['id' => 'customer_id']);
 	}
 
 	public function getOrderItems()
 	{
-		return $this->hasMany(OrderItem::className(), ['order_id' => ActiveRecord::PRIMARY_KEY_NAME]);
+		return $this->hasMany(OrderItem::className(), ['order_id' => 'id']);
 	}
 
 	public function getItems()
 	{
-		return $this->hasMany(Item::className(), [ActiveRecord::PRIMARY_KEY_NAME => 'item_id'])
+		return $this->hasMany(Item::className(), ['id' => 'item_id'])
 			->via('orderItems')->orderBy('id');
 	}
 
@@ -51,8 +57,8 @@ class Order extends ActiveRecord
 
 //	public function getBooks()
 //	{
-//		return $this->hasMany('Item', [ActiveRecord::PRIMARY_KEY_NAME => 'item_id'])
-//			->viaTable('tbl_order_item', ['order_id' => ActiveRecord::PRIMARY_KEY_NAME])
+//		return $this->hasMany('Item', ['id' => 'item_id'])
+//			->viaTable('tbl_order_item', ['order_id' => 'id'])
 //			->where(['category_id' => 1]);
 //	}
 
@@ -64,5 +70,25 @@ class Order extends ActiveRecord
 		} else {
 			return false;
 		}
+	}
+
+	/**
+	 * sets up the index for this record
+	 * @param Command $command
+	 */
+	public static function setUpMapping($command)
+	{
+		$command->deleteMapping(static::index(), static::type());
+		$command->setMapping(static::index(), static::type(), [
+			static::type() => [
+				"_id" => ["path" => "id", "index" => "not_analyzed", "store" => "yes"],
+				"properties" => [
+					"customer_id" => ["type" => "integer"],
+//					"create_time" => ["type" => "string", "index" => "not_analyzed"],
+					"total" => ["type" => "integer"],
+				]
+			]
+		]);
+
 	}
 }