+++ extensions/yii/elasticsearch/CHANGELOG.md
@@ -5,6 +5,7 @@ Yii Framework 2 elasticsearch extension Change Log
 ----------------------------
 
 - Enh #1382: Added a debug toolbar panel for elasticsearch (cebe)
+- Chg: Changed handling of ActiveRecord primary keys (cebe)
 
 2.0.0 alpha, December 1, 2013
 -----------------------------